Ethflashlogic

+++ contracts/interfaces/WrappedEtherInterface.sol
@@ -2,7 +2,9 @@
 pragma solidity ^0.6.8;
 
 interface WrappedEtherInterface {
-  function deposit() external payable;
-  function withdraw(uint amount) external;
-  function balanceOf(address owner) external view returns (uint256 balance);
+    function deposit() external payable;
+
+    function withdraw(uint256 amount) external;
+
+    function balanceOf(address owner) external view returns (uint256 balance);
 }
